7zip21.02 source code

+++ CPP/7zip/UI/FileManager/ExtractCallback.cpp
@@ -262,6 +262,7 @@ STDMETHODIMP CExtractCallbackImp::ShowMessage(const wchar_t *s)
 
 #endif
 
+void SetExtractErrorMessage(Int32 opRes, Int32 encrypted, const wchar_t *fileName, UString &s);
 void SetExtractErrorMessage(Int32 opRes, Int32 encrypted, const wchar_t *fileName, UString &s)
 {
   s.Empty();
@@ -335,20 +336,18 @@ void SetExtractErrorMessage(Int32 opRes, Int32 encrypted, const wchar_t *fileNam
       s += msg;
     else
     {
-      char temp[16];
-      ConvertUInt32ToString(opRes, temp);
-      s.AddAscii("Error #");
-      s.AddAscii(temp);
+      s += "Error #";
+      s.Add_UInt32(opRes);
     }
 
     if (encrypted && opRes != NArchive::NExtract::NOperationResult::kWrongPassword)
     {
-      // s.AddAscii(" : ");
+      // s += " : ";
       // AddLangString(s, IDS_EXTRACT_MSG_ENCRYPTED);
-      s.AddAscii(" : ");
+      s += " : ";
       AddLangString(s, IDS_EXTRACT_MSG_WRONG_PSW_GUESS);
     }
-    s.AddAscii(" : ");
+    s += " : ";
     s += fileName;
   }
 }
@@ -451,6 +450,7 @@ static void AddNewLineString(UString &s, const UString &m)
   s.Add_LF();
 }
 
+UString GetOpenArcErrorMessage(UInt32 errorFlags);
 UString GetOpenArcErrorMessage(UInt32 errorFlags)
 {
   UString s;
@@ -466,7 +466,7 @@ UString GetOpenArcErrorMessage(UInt32 errorFlags)
       continue;
     if (f == kpv_ErrorFlags_EncryptedHeadersError)
     {
-      m.AddAscii(" : ");
+      m += " : ";
       AddLangString(m, IDS_EXTRACT_MSG_WRONG_PSW_GUESS);
     }
     if (!s.IsEmpty())
@@ -483,7 +483,7 @@ UString GetOpenArcErrorMessage(UInt32 errorFlags)
     ConvertUInt32ToHex(errorFlags, sz + 2);
     if (!s.IsEmpty())
       s.Add_LF();
-    s.AddAscii(sz);
+    s += sz;
   }
   
   return s;
@@ -503,7 +503,7 @@ static void ErrorInfo_Print(UString &s, const CArcErrorInfo &er)
   if (warningFlags != 0)
   {
     s += GetNameOfProperty(kpidWarningFlags, L"Warnings");
-    s.AddAscii(":");
+    s += ":";
     s.Add_LF();
     AddNewLineString(s, GetOpenArcErrorMessage(warningFlags));
   }
@@ -511,7 +511,7 @@ static void ErrorInfo_Print(UString &s, const CArcErrorInfo &er)
   if (!er.WarningMessage.IsEmpty())
   {
     s += GetNameOfProperty(kpidWarning, L"Warning");
-    s.AddAscii(": ");
+    s += ": ";
     s += er.WarningMessage;
     s.Add_LF();
   }
@@ -519,12 +519,13 @@ static void ErrorInfo_Print(UString &s, const CArcErrorInfo &er)
 
 static UString GetBracedType(const wchar_t *type)
 {
-  UString s = L'[';
+  UString s ('[');
   s += type;
-  s += L']';
+  s += ']';
   return s;
 }
 
+void OpenResult_GUI(UString &s, const CCodecs *codecs, const CArchiveLink &arcLink, const wchar_t *name, HRESULT result);
 void OpenResult_GUI(UString &s, const CCodecs *codecs, const CArchiveLink &arcLink, const wchar_t *name, HRESULT result)
 {
   FOR_VECTOR (level, arcLink.Arcs)
@@ -634,7 +635,9 @@ HRESULT CExtractCallbackImp::ExtractResult(HRESULT result)
   if (result == S_OK)
     return result;
   NumArchiveErrors++;
-  if (result == E_ABORT || result == ERROR_DISK_FULL)
+  if (result == E_ABORT
+      || result == HRESULT_FROM_WIN32(ERROR_DISK_FULL)
+      )
     return result;
 
   Add_ArchiveName_Error();
@@ -704,7 +707,7 @@ STDMETHODIMP CExtractCallbackImp::AskWrite(
     {
       if (!destFileInfo.IsDir())
       {
-        RINOK(MessageError("can not replace file with folder with same name", destPathSys));
+        RINOK(MessageError("Cannot replace file with folder with same name", destPathSys));
         return E_ABORT;
       }
       *writeAnswer = BoolToInt(false);
@@ -713,7 +716,7 @@ STDMETHODIMP CExtractCallbackImp::AskWrite(
   
     if (destFileInfo.IsDir())
     {
-      RINOK(MessageError("can not replace folder with file with same name", destPathSys));
+      RINOK(MessageError("Cannot replace folder with file with same name", destPathSys));
       *writeAnswer = BoolToInt(false);
       return S_OK;
     }
@@ -727,7 +730,7 @@ STDMETHODIMP CExtractCallbackImp::AskWrite(
         Int32 overwriteResult;
         UString destPathSpec = destPath;
         int slashPos = destPathSpec.ReverseFind_PathSepar();
-        destPathSpec.DeleteFrom(slashPos + 1);
+        destPathSpec.DeleteFrom((unsigned)(slashPos + 1));
         destPathSpec += fs2us(destFileInfo.Name);
 
         RINOK(AskOverwrite(
@@ -748,24 +751,31 @@ STDMETHODIMP CExtractCallbackImp::AskWrite(
           default:
             return E_FAIL;
         }
+        break;
       }
+      default:
+        break;
     }
     
     if (OverwriteMode == NExtract::NOverwriteMode::kRename)
     {
       if (!AutoRenamePath(destPathSys))
       {
-        RINOK(MessageError("can not create name for file", destPathSys));
+        RINOK(MessageError("Cannot create name for file", destPathSys));
         return E_ABORT;
       }
       destPathResultTemp = fs2us(destPathSys);
     }
     else
+    {
+      if (NFind::DoesFileExist_Raw(destPathSys))
       if (!NDir::DeleteFileAlways(destPathSys))
+      if (GetLastError() != ERROR_FILE_NOT_FOUND)
       {
-        RINOK(MessageError("can not delete output file", destPathSys));
+        RINOK(MessageError("Cannot delete output file", destPathSys));
         return E_ABORT;
       }
+    }
   }
   *writeAnswer = BoolToInt(true);
   return StringToBstr(destPathResultTemp, destPathResult);
@@ -948,9 +958,8 @@ STDMETHODIMP CExtractCallbackImp::SetOperationResult7(Int32 opRes, Int32 encrypt
 }
 
 
-static const size_t k_SizeT_MAX = (size_t)((size_t)0 - 1);
 
-static const UInt32 kBlockSize = ((UInt32)1 << 31);
+// static const UInt32 kBlockSize = ((UInt32)1 << 31);
 
 STDMETHODIMP CVirtFileSystem::Write(const void *data, UInt32 size, UInt32 *processedSize)
 {
@@ -975,8 +984,9 @@ STDMETHODIMP CVirtFileSystem::Write(const void *data, UInt32 size, UInt32 *proce
       if (b < a)
         b = a;
       useMem = false;
-      if (b <= k_SizeT_MAX && b <= MaxTotalAllocSize)
-        useMem = file.Data.ReAlloc_KeepData((size_t)b, (size_t)file.Size);
+      const size_t b_sizet = (size_t)b;
+      if (b == b_sizet && b <= MaxTotalAllocSize)
+        useMem = file.Data.ReAlloc_KeepData(b_sizet, (size_t)file.Size);
     }
     if (useMem)
     {
@@ -1009,7 +1019,7 @@ HRESULT CVirtFileSystem::FlushToDisk(bool closeLast)
       {
         _outFileStream.Release();
         return E_FAIL;
-        // MessageBoxMyError(UString(L"Can't create file ") + fs2us(tempFilePath));
+        // MessageBoxMyError(UString("Can't create file ") + fs2us(tempFilePath));
       }
       _fileIsOpen = true;
       RINOK(WriteStream(_outFileStream, file.Data, (size_t)file.Size));
@@ -1027,7 +1037,7 @@ HRESULT CVirtFileSystem::FlushToDisk(bool closeLast)
     _numFlushed++;
     _fileIsOpen = false;
     if (file.AttribDefined)
-      NDir::SetFileAttrib(path, file.Attrib);
+      NDir::SetFileAttrib_PosixHighDetect(path, file.Attrib);
   }
   return S_OK;
 }
